Create a class for the Sampler
In a recent commit, the mechanism `shrinkctr` was replaced by a
sigma_min value. This makes the current code slightly different from
previous versions. In the old code, the `shrinkctr` would not decrease
when the value of sigma is doubled, which seems awkward. By comparing
current sigma and sigma_min, we can avoid this problem.