The black-box objective function should evaluate a list of points
This is done so that we can do batch sampling. Moreover, we aim to
use the library to find minima of expensive functions, so it may make
sense to evaluate the function at multiple points at once. By doing so,
we also pass the responsibility of parallelizing the function to the
user. As a consequence, we do not measure the time it takes to evaluate
the function anymore.