Fix issue with SLHD for integers.
It was possible to get repeated points in the SLHD for integers if the
number of points was equal to the number of possible values minus one.
This was due to the rounding strategy used to convert the continuous
values to integers.

Moreover, this commit fix a bug in the SLHD.