Use the min value of the GP in `bayesian_optimization()`.
This is based o feedback from Robi Vercellino for
problems with shallow global minimum. MinEI may avoid good regions.
That happens because EI is based on the trained model and not on the training data.
Training data may be missing at a local minimum, but GP may already have a good guess there,
so minEI will never sample there, and we will never find out the true value of the objective function.